Refactor statscollectors module
Split stats collectors custom classes in separated modules, so the code
is a bit more organized and easier to follow.

During the split, we renamed DashCollectionsStatsHistoryCollector to
ScrapyCloudCollectionsStatsHistoryCollector which is a name that
indicates better where the stats will be stored. As this class wasn't
delivered in any Spidermon release, we could rename it without having
backward problems.

Moved LocalStorageStatsHistoryCollector to a new module, but because
this was release in old versions of Spidermon, we needed to ensure that
the old import location
(spidermon.contrib.stats.statscollectors.LocalStorageStatsHistoryCollector)
is still acessible and won't break any code that is already using this
stats collector.
