+++ spidermon/contrib/stats/statscollectors/__init__.py
@@ -1,0 +1,3 @@
+from spidermon.contrib.stats.statscollectors.local_storage import (
+    LocalStorageStatsHistoryCollector,
+)
